Profiling metal backend(3)
