+++ tool/profile_visualizer/__init__.py
@@ -1,0 +1,5 @@
+"""Profiling visualizer package."""
+
+from .server import start_server
+
+__all__ = ["start_server"]
