+++ tool/profile_visualizer/static/index.html
@@ -1,0 +1,73 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>Profiling Visualizer</title>
+    <link rel="stylesheet" href="styles.css">
+</head>
+<body>
+    <div class="container">
+        <div class="header">
+            <h1>🔥 Profiling Visualizer</h1>
+            <div class="meta" id="metadata"></div>
+        </div>
+
+        <div class="file-loader" id="fileLoader">
+            <h3>📂 Load Profiling Data</h3>
+
+            <div class="file-selector">
+                <div class="directory-browser" id="directoryBrowser">
+                    <div class="loading-spinner"></div>
+                    <div class="current-path" id="currentPath">Loading...</div>
+                    <ul class="dir-list" id="dirList"></ul>
+                    <ul class="file-list" id="fileList"></ul>
+                </div>
+                <button class="load-button" id="loadButton" disabled onclick="loadSelectedFile()">Load Selected File</button>
+            </div>
+
+            <div class="separator">OR</div>
+
+            <div class="drop-zone" id="dropZone">
+                <input type="file" id="fileInput" accept=".json" style="display: none;">
+                <div class="drop-zone-text">📁 Drop JSON file here or click to browse</div>
+                <div class="drop-zone-hint">Supports local JSON files</div>
+            </div>
+        </div>
+
+        <div id="contentWrapper" class="content-wrapper">
+            <div class="stats" id="stats"></div>
+
+            <div class="tab-buttons">
+                <button class="tab-button active" onclick="switchTab('tree')">Tree View</button>
+                <button class="tab-button" onclick="switchTab('table')">Table View</button>
+                <button class="tab-button" onclick="switchTab('flame')">Flame Graph</button>
+            </div>
+
+            <div class="controls">
+                <label>
+                    Min Time (ms):
+                    <input type="number" id="minTime" value="0" step="0.1" onchange="applyFilters()">
+                </label>
+                <label>
+                    <input type="checkbox" id="expandAll" onchange="toggleExpandAll()">
+                    Expand All
+                </label>
+            </div>
+
+            <div id="tree-tab" class="tab-content active">
+                <div class="tree-container" id="treeView"></div>
+            </div>
+
+            <div id="table-tab" class="tab-content">
+                <div class="tree-container" id="tableView"></div>
+            </div>
+
+            <div id="flame-tab" class="tab-content">
+                <div class="flamegraph" id="flameView"></div>
+            </div>
+        </div>
+    </div>
+
+    <script src="app.js"></script>
+</body>
+</html>
